Move ExtensibleObject to CesiumUtility
