+++ CesiumJsonReader/include/CesiumJsonReader/IExtensionJsonHandler.h
@@ -1,19 +1,17 @@
 #pragma once
 
-#include "CesiumJsonReader/ExtensibleObject.h"
 #include "CesiumJsonReader/IJsonHandler.h"
+#include <CesiumUtility/ExtensibleObject.h>
 #include <any>
 #include <string_view>
 
 namespace CesiumJsonReader {
 
-struct ExtensibleObject;
-
 class IExtensionJsonHandler : public IJsonHandler {
 public:
   virtual void reset(
       IJsonHandler* pParentHandler,
-      ExtensibleObject& o,
+      CesiumUtility::ExtensibleObject& o,
       const std::string_view& extensionName) = 0;
 };
